+++ cli/pkg/workspace_repo/workspace_repo.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/chanzuckerberg/happy/cli/pkg/options"
-	"github.com/chanzuckerberg/happy/cli/pkg/util"
+	"github.com/chanzuckerberg/happy/shared/util"
 	"github.com/hashicorp/go-tfe"
 )